Moving `ast.defaultRegoVersion`->`internal/rego_default.DefaultRegoVersion`
Note: can't use ast.RegoVersion values directly because of circular dependencies.
`DefaultRegoVersion` could alternatively be set to int-representation (manifest representation), but the required call to `ast.RegoVersionFromInt()` from `ast.DefaultRegoVersion()` is likely somewhat more costly.
