--- ast/rego_default_v0.go
@@ -1,9 +1,0 @@
-// Copyright 2024 The OPA Authors.  All rights reserved.
-// Use of this source code is governed by an Apache2
-// license that can be found in the LICENSE file.
-
-//go:build !opa_rego_v1
-
-package ast
-
-var defaultRegoVersion = RegoV0
