+++ internal/rego_default/rego_default_v1.go
@@ -1,0 +1,11 @@
+// Copyright 2024 The OPA Authors.  All rights reserved.
+// Use of this source code is governed by an Apache2
+// license that can be found in the LICENSE file.
+
+//go:build opa_rego_v1
+// +build opa_rego_v1
+
+package rego_default
+
+// DefaultRegoVersion is the default Rego version (v1).
+var DefaultRegoVersion = 3 // enum value for ast.RegoV1
